+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "next": "12.1.0",
+    "prettier": "2.5.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "sass": "^1.49.9"
