+++ pages/[id].tsx
@@ -1,14 +1,14 @@
 /*나중에는 [id] 이런 식으로 바꾸면 될 듯 */
 
-import React, {useEffect} from 'react';
-import {useRouter} from "next/router";
-import Profile from "../components/Body/Profile";
+import React, { useEffect } from "react";
+import { useRouter } from "next/router";
+import ProfileComponent from "../components/Body/ProfileComponent";
 
 const Id = () => {
-    const router = useRouter();
-    const id = Number(router.query.id)
+  const router = useRouter();
+  const id = Number(router.query.id);
 
-    return <Profile id={id}/>
+  return <ProfileComponent id={id} />;
 };
-Id.displayName = "Profile";
+Id.displayName = "ProfileComponent";
 export default Id;
