+++ pages/_app.tsx
@@ -1,27 +1,23 @@
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-import DefaultLayout from "../components/Layout/DefaultLayout";
-import OnlyHeaderLayout from "../components/Layout/OnlyHeaderLayout";
+import "../styles/globals.css";
+import type { AppProps } from "next/app";
+import DefaultLayoutComponent from "../components/Layout/DefaultLayoutComponent";
+import OnlyHeaderLayoutComponent from "../components/Layout/OnlyHeaderLayoutComponent";
 
 function MyApp({ Component, pageProps }: AppProps) {
-    switch (Component.displayName) {
-        case "Profile":
-            return (
-                <OnlyHeaderLayout>
-                    <Component {...pageProps} />
-                </OnlyHeaderLayout>
-            )
-        default:
-            return (
-                <DefaultLayout>
-                    <Component {...pageProps} />
-                </DefaultLayout>
-            )
-    }
-
-
-
-
+  switch (Component.displayName) {
+    case "Profile":
+      return (
+        <OnlyHeaderLayoutComponent>
+          <Component {...pageProps} />
+        </OnlyHeaderLayoutComponent>
+      );
+    default:
+      return (
+        <DefaultLayoutComponent>
+          <Component {...pageProps} />
+        </DefaultLayoutComponent>
+      );
+  }
 }
 
-export default MyApp
+export default MyApp;
