+++ styles/Header1.module.scss
@@ -5,7 +5,7 @@
   align-items: center;
   padding-left: 30px;
   padding-right: 30px;
-  border-bottom: 1px solid #F2F2F2;
+  border-bottom: 1px solid #f2f2f2;
 }
 
 .headerText {
@@ -15,16 +15,15 @@
   font-size: 25px;
   line-height: 31px;
   text-align: center;
-  color: #588A69;
-
+  color: #588a69;
 }
 
 .searchLabel {
   display: flex;
   width: 718px;
   height: 38px;
-  background: #F3F3F3;
-  border: 4px solid #588A69;
+  background: #f3f3f3;
+  border: 4px solid #588a69;
   box-sizing: border-box;
   align-items: center;
   justify-content: right;
@@ -35,7 +34,7 @@
   height: 100%;
   width: 100%;
   border: none;
-  background-color: #F3F3F3;
+  background-color: #f3f3f3;
   font-size: 18px;
   font-family: Nanum Gothic;
   font-style: normal;
@@ -55,7 +54,7 @@
   width: 38px;
   height: 38px;
   border-radius: 9999px;
-  background-color: #C4C4C4;
+  background-color: #c4c4c4;
   display: flex;
   justify-content: center;
   align-items: center;
