+++ styles/Header2.module.scss
@@ -1,7 +1,7 @@
 @mixin button_default {
   width: 80px;
   height: 40px;
-  background-color: #FFFFFF;
+  background-color: #ffffff;
   border: none;
   justify-content: center;
   font-family: Nanum Gothic;
@@ -21,12 +21,11 @@
 .header2 {
   display: flex;
   height: 50px;
-  width:100%;
+  width: 100%;
   justify-content: space-between;
   align-items: center;
-  border-bottom: 1px solid #F2F2F2;
+  border-bottom: 1px solid #f2f2f2;
   padding-right: 30px;
-
 }
 .buttonGroup {
   display: flex;
@@ -39,6 +38,6 @@
   width: 120px;
   height: 35px;
   color: white;
-  background: #EA910C;
+  background: #ea910c;
   font-size: 18px;
 }
