+++ styles/Layout.module.scss
@@ -18,11 +18,9 @@
   display: flex;
 }
 .pageMain {
-
   width: calc(100% - 300px);
 }
 .pageSideBar {
-  border-left: 1px solid #F2F2F2;
+  border-left: 1px solid #f2f2f2;
   width: 300px;
 }
-
