+++ styles/MainPage.module.scss
@@ -9,5 +9,5 @@
   text-align: center;
   padding-left: 30px;
   height: 80px;
-  border-bottom: 1px solid #F2F2F2;
+  border-bottom: 1px solid #f2f2f2;
 }
