+++ styles/OnlyHeaderLayout.module.scss
@@ -15,6 +15,3 @@
 .pageContent {
   width: 100%;
 }
-
-
-
