+++ styles/SideComponent.module.scss
@@ -17,5 +17,5 @@
 }
 .sideComponentContent {
   height: 300px;
-  background-color: #E5EAD9;
+  background-color: #e5ead9;
 }
