+++ styles/SimplifiedPreview.module.scss
@@ -39,17 +39,17 @@
 }
 .previewHeaderTag1 {
   @include tag;
-  background-color: #588A69;
+  background-color: #588a69;
 }
 .previewHeaderTag2 {
   @include tag;
-  background-color: #68A37C;
+  background-color: #68a37c;
 }
 .previewHeaderTag3 {
   @include tag;
-  background-color: #79BA8F;
+  background-color: #79ba8f;
 }
-.previewHeader2{
+.previewHeader2 {
   display: flex;
   align-items: center;
   height: 50px;
@@ -60,9 +60,8 @@
   line-height: 23px;
   letter-spacing: 0em;
   text-align: left;
-  color: #588A69;
+  color: #588a69;
 }
 .previewBody {
   margin-bottom: 10px;
 }
-
