+++ styles/TagStatBox.module.scss
@@ -3,7 +3,7 @@
   height: 80px;
   display: flex;
   justify-content: center;
-  border: 1px solid #588A69;
+  border: 1px solid #588a69;
   font-family: NanumGothic;
   font-style: normal;
   font-weight: normal;
@@ -13,5 +13,5 @@
   align-items: center;
   border-radius: 5px;
   text-align: center;
-  color: #35523F;
+  color: #35523f;
 }
